chore(livePage/render): 필요없는 파일 삭제 및 코드 제거
Resolves: #162
