--- src/app/(route)/live/[host_uid]/components/LiveDetails/LiveDetailsOptionButton.client.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import SvgIcon from '@/app/_components/SVGIcon.server'
-
-//모달 추가해야함..
-const LiveOptionButton = () => {
-  return (
-    <button className="bg-gray-100 rounded-[17px] px-2 h-[34px] inline-flex items-center justify-center gap-x-[4px] text-sm leading-none font-sans cursor-pointer transition-all duration-200 ease-in-out relative align-top focus:outline-none">
-        <SvgIcon name="VideoDots" width={16} height={4}/>
-    </button>
-  );
-};
-
-export default LiveOptionButton;
